// js/stat-evaluation-player/src/generated/PassPlayerStats.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type PassPlayerStats = { completed_pass_count: number, received_pass_count: number, total_pass_distance: number, total_pass_advance: number, longest_pass_distance: number, is_last_completed_pass: boolean, last_completed_pass_time: number | null, last_completed_pass_frame: number | null, time_since_last_completed_pass: number | null, frames_since_last_completed_pass: number | null, };