// js/stat-evaluation-player/src/generated/PassTeamStats.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Per-team accumulated pass stats.
 */
export type PassTeamStats = { completed_pass_count: number, total_pass_distance: number, total_pass_advance: number, longest_pass_distance: number, };